Tsumino: Refactor
